Refactor server package: split monolith, fix error handling, add helper functions
Split the 1471-line server.go monolith into domain-specific files:
- movie_handlers.go, series_handlers.go, search_handlers.go
- indexer_handlers.go, quality_handlers.go, download_handlers.go
- job_handlers.go, static_handlers.go, routes.go, helpers.go

Key changes:

- Add request helpers (parseURLInt64, decodeJSON, respond, respondError)
  that ensure consistent JSON error responses across all ~50 handlers,
  replacing mixed http.Error (text/plain) and writeErrorResponse (JSON)

- Fix Serve() to propagate bind failures immediately via channel-based
  select instead of blocking on SIGINT indefinitely as a zombie process

- Fix DELETE /indexers and DELETE /quality/definitions to read ID from
  URL path instead of request body (consistent with other DELETE endpoints)

- Fix search handlers returning 400 for internal errors (now 500)

- Add 404 handling for GetMovieDetailByTMDBID and GetTVDetailByTMDBID
  (was returning 500 for not-found)

- Remove raw request body from error responses in decodeJSON (security)

- Restore request-scoped logging in response helpers (regression fix:
  helpers now accept *http.Request and use the context-scoped logger
  with path and request ID from LogMiddleware)

- Split server_test.go into domain-aligned _test.go files matching
  their handler counterparts with minimal per-file imports

- Update frontend API calls for DELETE endpoint changes
